FIXED: apart from touches and CALayer auto-generation (the SVGKImage.CALayerTree property is never set at the moment), this is now back to working status - but with SVGKImage as a complete replacement for the old SVGDocument.
NB: the DOM classes included here are about 30% non-implemented, and are NOT integrated into the parser.

NB: the SVG-DOM classes have not in ANY WAY been fixed to make them adhere to the spec - so I have left them in a different sub-folder. As they are "fixed", they should be moved to the "SVG-DOM" folder, one by one.